[ADF-3824] Reorganize Task Cloud Folder
