+++ lib/process-services-cloud/src/lib/process-services-cloud.module.ts
@@ -18,7 +18,6 @@
 import { NgModule } from '@angular/core';
 import { TRANSLATION_PROVIDER } from '@alfresco/adf-core';
 import { AppListCloudModule } from './app-list-cloud/app-list-cloud.module';
-import { TaskListCloudModule } from './task-list-cloud/task-list-cloud.module';
 import { TaskCloudModule } from './task-cloud/task-cloud.module';
 import { ProcessListCloudModule } from './process-list-cloud/process-list-cloud.module';
 import { ProcessCloudModule } from './process-cloud/process-cloud.module';
@@ -27,7 +26,6 @@ import { StartTaskCloudModule } from './start-task-cloud/start-task-cloud.module
 @NgModule({
     imports: [
         AppListCloudModule,
-        TaskListCloudModule,
         TaskCloudModule,
         ProcessListCloudModule,
         ProcessCloudModule,
@@ -45,7 +43,6 @@ import { StartTaskCloudModule } from './start-task-cloud/start-task-cloud.module
     ],
     exports: [
         AppListCloudModule,
-        TaskListCloudModule,
         TaskCloudModule,
         ProcessListCloudModule,
         ProcessCloudModule,
