+++ lib/process-services-cloud/src/lib/styles/_index.scss
@@ -1,6 +1,6 @@
 @import './../app-list-cloud/components/app-details-cloud.component';
 @import './../app-list-cloud/components/app-list-cloud.component';
-@import './../task-cloud/task-filters-cloud/task-filters-cloud.component.scss';
+@import './../task-cloud/components/task-filters-cloud.component.scss';
 @import './../process-list-cloud/components/process-list-cloud.component.scss';
 @import './../start-task-cloud/components/start-task-cloud.component';
 @import './../start-task-cloud/components/people-cloud/people-cloud.component';
