+++ lib/process-services-cloud/src/lib/task-cloud/models/filter-cloud.model.ts
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+import { TaskListCloudSortingModel } from './task-list-sorting.model';
+
 export class TaskFilterCloudModel  {
     id: string;
     name: string;
@@ -48,4 +50,55 @@ export class TaskFilterCloudModel  {
 export interface FilterActionType {
     actionType: string;
     id: string;
+}
+export class TaskQueryCloudRequestModel {
+    appName: string;
+    appVersion?: string;
+    assignee?: string;
+    claimedDate?: string;
+    createdDate?: Date;
+    description?: string;
+    dueDate?: null;
+    id?: string;
+    name?: string;
+    owner?: string;
+    parentTaskId?: string;
+    priority?: number;
+    processDefinitionId?: string;
+    processInstanceId?: string;
+    serviceFullName?: string;
+    serviceName?: string;
+    serviceType?: string;
+    serviceVersion?: string;
+    status?: string;
+    maxItems: number;
+    skipCount: number;
+    sorting?: TaskListCloudSortingModel[];
+
+    constructor(obj?: any) {
+        if (obj) {
+            this.appName = obj.appName;
+            this.appVersion = obj.appVersion;
+            this.assignee = obj.assignee;
+            this.claimedDate = obj.claimedDate;
+            this.createdDate = obj.createdDate;
+            this.description = obj.description;
+            this.dueDate = obj.dueDate;
+            this.id = obj.id;
+            this.name = obj.name;
+            this.owner = obj.owner;
+            this.parentTaskId = obj.parentTaskId;
+            this.priority = obj.priority;
+            this.processDefinitionId = obj.processDefinitionId;
+            this.processInstanceId = obj.processInstanceId;
+            this.serviceFullName = obj.serviceFullName;
+            this.serviceName = obj.serviceName;
+            this.serviceType = obj.serviceType;
+            this.serviceVersion = obj.serviceVersion;
+            this.status = obj.status;
+            this.maxItems = obj.maxItems;
+            this.skipCount = obj.skipCount;
+            this.sorting = obj.sorting;
+        }
+    }
 }
