+++ lib/process-services-cloud/src/lib/task-cloud/services/task-list-cloud.service.spec.ts
@@ -20,14 +20,14 @@ import { setupTestBed } from '@alfresco/adf-core';
 import { fakeTaskCloudList } from '../mock/fakeTaskResponseMock';
 import { AlfrescoApiServiceMock, LogService, AppConfigService, StorageService, CoreModule } from '@alfresco/adf-core';
 import { TaskListCloudService } from './task-list-cloud.service';
-import { TaskQueryCloudRequestModel } from '../models/filter-cloud-model';
+import { TaskQueryCloudRequestModel } from '../models/filter-cloud.model';
 
 describe('Activiti TaskList Cloud Service', () => {
 
     let service: TaskListCloudService;
     let alfrescoApiMock: AlfrescoApiServiceMock;
 
-    function returFakeTaskListResults() {
+    function returnFakeTaskListResults() {
         return {
             oauth2Auth: {
                 callCustomApi : () => {
@@ -72,7 +72,7 @@ describe('Activiti TaskList Cloud Service', () => {
 
     it('should return the tasks', (done) => {
         let taskRequest: TaskQueryCloudRequestModel = <TaskQueryCloudRequestModel> { appName: 'fakeName' };
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returFakeTaskListResults);
+        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskListResults);
         service.getTaskByRequest(taskRequest).subscribe((res) => {
             expect(res).toBeDefined();
             expect(res).not.toBeNull();
