+++ lib/process-services-cloud/src/lib/task-cloud/task-cloud.module.ts
@@ -19,14 +19,16 @@ import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { TaskFiltersCloudComponent } from './task-filters-cloud/task-filters-cloud.component';
+import { TaskFiltersCloudComponent } from './components/task-filters-cloud.component';
 import { MaterialModule } from '../material.module';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
-import { TranslateLoaderService, LogService, StorageService } from '@alfresco/adf-core';
+import { TranslateLoaderService, LogService, StorageService, TemplateModule, DataTableModule } from '@alfresco/adf-core';
 import { TaskFilterCloudService } from './services/task-filter-cloud.service';
 import { HttpClientModule } from '@angular/common/http';
-import { EditTaskFilterCloudComponent } from './task-filters-cloud/edit-task-filter-cloud.component';
-import { TaskFilterDialogCloudComponent } from './task-filters-cloud/task-filter-dialog-cloud.component';
+import { EditTaskFilterCloudComponent } from './components/edit-task-filter-cloud.component';
+import { TaskFilterDialogCloudComponent } from './components/task-filter-dialog-cloud.component';
+import { TaskListCloudComponent } from './components/task-list-cloud.component';
+import { TaskListCloudService } from './services/task-list-cloud.service';
 @NgModule({
     imports: [
         FormsModule,
@@ -40,11 +42,27 @@ import { TaskFilterDialogCloudComponent } from './task-filters-cloud/task-filter
                 useClass: TranslateLoaderService
             }
         }),
-        MaterialModule
+        MaterialModule,
+        TemplateModule,
+        DataTableModule
     ],
-    declarations: [TaskFiltersCloudComponent, EditTaskFilterCloudComponent, TaskFilterDialogCloudComponent],
-    exports: [TaskFiltersCloudComponent, EditTaskFilterCloudComponent],
-    providers: [TaskFilterCloudService, LogService, StorageService],
+    declarations: [
+        TaskFiltersCloudComponent,
+        EditTaskFilterCloudComponent,
+        TaskFilterDialogCloudComponent,
+        TaskListCloudComponent
+    ],
+    exports: [
+        TaskFiltersCloudComponent,
+        EditTaskFilterCloudComponent,
+        TaskListCloudComponent
+    ],
+    providers: [
+        TaskFilterCloudService,
+        TaskListCloudService,
+        LogService,
+        StorageService
+    ],
     entryComponents: [
         TaskFilterDialogCloudComponent
     ]
